tag service

+++ blog-service/pkg/app/app.go
@@ -15,7 +15,7 @@ type Response struct {
 type Pager struct {
 	Page      int `json:"page"`
 	PageSize  int `json:"page_size"`
-	TotalRows int `json:"total_rows"`
+	TotalRows int64 `json:"total_rows"`
 }
 
 func NewResponse(ctx *gin.Context) *Response {
@@ -29,7 +29,7 @@ func (r *Response) ToResponse(data interface{}) {
 	r.Ctx.JSON(http.StatusOK, data)
 }
 
-func (r *Response) ToResponseList(list interface{}, totalRows int) {
+func (r *Response) ToResponseList(list interface{}, totalRows int64) {
 	r.Ctx.JSON(http.StatusOK, gin.H{
 		"list": list,
 		"pager": Pager{
